# Optimize precomputing of websocket event